refactor: move give item to helper
